Utilizing sb-admin template
I created this branch to provide a better foundation for us to move forward.  I would recommend creating a copy of "index.html" and then modifying the heck out of it making it whatever you want to be.

In short, make a copy of index.html, and strip away everything you don't need.