CLEAN FRONT END
I have cleaned up the front end and HEAVILY documented js/controllers/MainController.js as well as views/admin/admin-prospects.html for reference on how to use angular.js.

Lab 8 is also an excellent reference, but in this project we now have examples of going to and from the database.