Stable single page app functionality
I just finished making the site a single page app. The way you need to reference URLS  is a little different now.
http://oraserv.cs.siena.edu/~perm_team1_2017/nick/#/     <- This is the homepage
http://oraserv.cs.siena.edu/~perm_team1_2017/nick/#/admin <- This brings you to admin.html
http://oraserv.cs.siena.edu/~perm_team1_2017/nick/#/home <- This brings you to user.html
The key for where each of these URL keywords resolves to can be found in `project/js/app.js`
Any new page added needs to have its route added to that file (just follow the other examples, its really easy. And then the navigation links would need to be updated as well. but now instead of putting `admin.html` in an href, you would write `#/admin` in the href. Any questions feel free to text me.
I am sending a pull request to @bebber, he's going to review it and merge it in when he sees fit.